[Linux] Fixes search engines not persisting for locale. (#30319)
* [Linux] Fixes search engines not persisting for locale.

On Linux, unlike other platforms, RegionalCapabilitiesServiceClientLinux::FetchCountryId
only relies on variations service to retrieve the locale. On the initial run, there's no
valid locale set from the variations service and the regional capabilities service falls
back on the device locale. On the consecutive runs the locale is retrieved from the
variations service and causes the search engines selection to change as well. This makes
little sense since the user would likely prefer to retain the same set of search engines
regardless of their physical location.

This change overrides RegionalCapabilitiesServiceClientLinux::FetchCountryId to use the
platform-neutral implementation in RegionalCapabilitiesServiceClient that gets the locale
from the device.

* Disable failing upstream tests.
* Add a browser test to check capabilities service uses device locale.