Disable `brave:all` due to OoM on the Windows CI
This change disables `brave:all` builds in our infra, as we are running
into out-of-memory issues at the final linking step. This has to be
sorted through the use of `concurrent_links`, however we need some time
to figure out the best way to deploy this gn arg and the best value for
it.