Fix Leo right-click actions not working in PDF (uplift to 1.82.x) (#30404)
Fix Leo right-click actions not working in PDF (#30357)

Currently when trying to do right-click actions in PDF, it would fail
silently because it can't get a tab helper due to tab_helper won't be
installed in guest WebContents (PDF plugin). embedder_web_contents
should be used instead to access the tab helper, which is done in this
PR.

This PR also makes a change to not try rewrite-in-place for PDF case.
Current impl doesn't work for PDF since web_contents.Replace doesn't
work, so we should just go for regular panel operation in this case for
now.

Also, when error happens, we currently try to open AI chat panel using
source_web_contents_, which is incorrect, it should be
embedder_web_contents_. However, in this PR, we just won't go into
rewrite-in-place code path for PDF cases, hence no need to change
source_web_contents_ to embedder_web_contents_ inside this path.

Test coverage are also improved for both PDF and non-PDF cases.