[build] Support `concurrent_links` in `.env` (#30417)
This gn arg allows the control of how linking instances can happen
concurrently. This is necessary for the Windows PR build CI which is
running out of memory when linking targets.

Additionally, this change corrects the parsing of `.env` values to carry
over their type through, as `concurrent_links` is an integer. This has
affected a few other values that are set by CI as integers.

Resolves https://github.com/brave/brave-browser/issues/48090