[canary] `gnrt` support to B🚀 (#30439)
This PR adds support for `brockit lift` and `brockit regen` to support
`gnrt` reruns committing. With `lift`, a `gnrt` rerun occurs at the end
of the lift process, and if successful, all files changed under
`third_party/rust` are committed as rerun updates.

There are also changes to `regen` to support `gnrt` reruns, however a
new flag has been added to `regen`, namely, `--dry-run` to permit
running all regen operations without committing them. This is useful if
in the future we decide to hook up dry run checks for these process in
CI.

Resolves https://github.com/brave/brave-browser/issues/48140