[ubsan] Fix `TestingBrowserProcess` cast override
This override has been broken for a while, and has been doing a broken
cast that yields undefined behaviour, as the type is not the same as the
one in the downcast. This has never been noticed until we started
running UBsan.

This fix corrects this override, by using a patch in the original class
to instantiate the override class.

https://github.com/brave/internal/issues/1406