AI Chat: Create a semi-built-in Profile for agentic features (#30285)
* AI Chat: Create a semi-built-in Profile for agentic features

For now, these features will only be useable in this profile. The purpose is for users and team members to experiment with these features, isolated from their regular browser context with existing site authentication. Whilst these features might evenutally be usable in a user's main profile under isolated storage (a built-in Container), this allows us to decouple these features from that behavior being implemented, as well as try out the separate Profile experience.

- Flag for whether the agentic behavior and the separate profile is enabled
- UI for the button in the AI Chat WebUI that will launch the profile
- Profile creation params, theme, and behavior:
  - Doesn't cause the profile picker to be shown on browser startup

For subsequent PRs:
- A separate and basic NTP with messaging
- Content agent features