[bots] Move `gn` args to static `.gni` files (#30303)
This PR introduces:

  - `//brave/build/args/brave_defaults.gni`
  - `//brave/build/args/blink_platform_defaults.gni`

These two `.gni` sources are to be used to provide absolute overrides
for `gn` args that are always overridden in Brave builds. Unfortunately,
having a single `gni` was not possible, as `iOS` in particular does not
uses a good number of flags that the other platforms use, so
the blink platform defaults is provided for args that are supported in all
platforms but ios.

This change is one of many steps being taken to move handling of `gn`
args out of `config.js`.

Resolves https://github.com/brave/brave-browser/issues/47956