Leo image paste from clipboard support (#30540)
* Move convertFileToUploadedFile to file_utils

Extract the file conversion utility function from drag_overlay to a
reusable utility module for use across multiple components.

* Rename processDroppedImages to attachImages

Rename the function to be more generic as it will be used for both
drag-and-drop and clipboard paste operations.

* Support clipboard pasting

Add support for pasting images from clipboard in the input box.
Images are filtered and processed using the same attachImages
function used for drag and drop.

* Add test for pasting