[codehealth][CodeHealth] Simplify override for `TestingBrowserProcess`
The original override class, which uses inheritance has had to go
through a few fixes, as it riddled with Undefined Behaviour, however
once it became apparent patching was needed, the whole point of using
inheritance for this override is lost, as the current override is
bloated, and it is not avoiding patching in any way.

This PR simplifies the override for this type by having macro hooks
patched into the class, to allow us to manage the lifetime of
`TestingBraveBrowserProcess` through `TestingBraveBrowserProcess`.

Resolves https://github.com/brave/brave-browser/issues/48183