[iOS] Support the new obscured insets API in iOS 26
This change adds support for setting the obscured insets of the page which fixes the page always getting a fixed viewport. As a side effect, this now allows the web view to fill the entire space and let us use blur effects on the toolbars.

On iOS 18 and below the viewports will still be incorrect but we disable the web view's clipping so that it renders under the toolbars and maintain the same blur effects.