Apply pending updates on macOS when idle
This commit adds a Griffin feature that restarts the browser when there
is a pending update on macOS, no windows are open, clear data on exit
isn't set and the system has been idle for 3 minutes. The motivation is
to roll out updates on macOS more quickly. The feature is
default-disabled and meant to be enabled on the server on Nightly only
for now.