[cr141] Making `GetAllBraveDeviceInfo` an empty virtual
`DeviceInfoTracker::GetAllBraveDeviceInfo` was a pure virtual, however
this forces the definition of it to be provided everywhere. Recently a
mock class, namely `MockDeviceInfoTracker` was introduced in a test
file, and only patching can correct the overridding for this function.

This change defines a completely empty base definition, which permits
`MockDeviceInfoTracker` to not be concerned with `GetAllBraveDeviceInfo`
if not required.

Chromium changes:
https://chromium.googlesource.com/chromium/src/+/8daddb32db4c820c31d8aff4c59c6d9a22962b99

commit 8daddb32db4c820c31d8aff4c59c6d9a22962b99
Author: Teresa Mao <temao@chromium.org>
Date:   Wed Aug 20 10:28:25 2025 -0700

    [TabGroupsModule] Add device name for group if available.

    Fixed: 438289723
    Change-Id: I391aaa6077ce4df2f5766bc27bc7382195e060b8
    Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6841727
    Commit-Queue: Teresa Mao <temao@chromium.org>
    Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
    Reviewed-by: Elly FJ <ellyjones@chromium.org>
    Cr-Commit-Position: refs/heads/main@{#1504058}