Fixed web contents area is not clickable (#30742)
Resolves brave/brave-browser#48574

On macOS, ContentsBackground view controls consumed mouse event.
Fixed by preventing event handling from it.
As this view is used by rounded corners feature, this bug doesn't happen if it's diabled.
If contents view(or split view or multi contents view) is positioned on top of it, this bug doesn't happen.

TEST=BraveBrowserViewWithRoundedCornersTest.ContentsBackgroundEventHandleTest