Fixes PageInfoCookiesContentView override and tests.
https://github.com/brave/brave-core/pull/30773 changed the code to
remove the views instead of hiding them, but the tests didn't get
updated because they weren't running at the time.

Buto also, the current override crushes (only in Release) due to the
label being added as a child to a wrapper, so when we remove the
label we should remove the wrapper instead.