[AI Chat] Introduce ToolProvider and pass instances to ConversationHandler (#30724)
* [AI Chat] Introduce ToolFactory and pass instances to ConversationHandler

Creates tool factories that are passed from the browser to the ConversationHandler via the AIChatService as well as from the AIChatService itself.

A minimal stub so that we can parallel some branches of work:
- Tab management Tools
- Content agent Tools
- Basic stateful conversation Tools, e.g. TODO tool, AssistantDetailStorage tool (i.e. created and owned by the ConversationHandler, but have a lifecycle)

## `ToolProviderFactory`
Introduces `ToolProviderFactory` and `ToolProvider` concept so that different layers can provide Tools and we can control the ownership and lifecycle of them. There are two layers here so that a ToolProvider can maintain state across different tools per-Conversation.

## `BrowserToolProviderFactory`
The brave/browser layer provides a `BrowserToolProviderFactory` to `AIChatService` which uses it to create a `BrowserToolProvider` for each ConversationHandler.
This is important so that the `BrowserToolProvider` can maintain state across its group of Tool instances and not just per-Tab. For example, the content agent tools are a group of many tools that share state such as which Tabs are being acted upon.

## `ConversationToolProvider`
The use-case here is `components/aichat`-provided core tools, probably without shared state between them. So far it's only providing a single singleton tool so this will evolve.