[AI Chat] Global side panel for agent profile, or everywhere via feature flag (#30720)
This begins a much-requested feature to have the side panel be global to the window and not per-Tab, meaning

- The side panel stays open when switching active Tabs and not only for the Tab it was initially opened ageainst.
- The conversation does not change when navigating the active Tab or when switching to a different Tab.

It is enabled always for the AI Chat agent profile, which is the driving force behind implementing this change now. It is difficult to have a content agent conversation when navigating the content switches to a new conversation. And when implementing this change, I did not want to hardcode it only for agent profiles when I know this is also requested for the regular profiles.