Add memory storage tool for adding memories in Leo UI during conversations (#30660)
This PR adds a memory storage tool, which will be provided to agent if the model supports tools.
When agent request to use this tool via tool_use event, a memory string parameter should be provided and we will run UseTool to add the memory and reply with an empty text response if success or an error message.
The added memory will be displayed in chat UI for user with options to undo or manage all memories via settings page link.
This tool is currently owned by AIChatService and is only available when memory feature preference is enabled.

In the first iteration we are starting with a limited scope (we instruct agents to do below via our tool/input descriptions):
- It would only be triggered when user makes an explicit request like "Remember ..." or similar commands.
- It currently only add new memory entries and won't try to edit existing memories.
- Only expect to be called once by agent for each turn.
- Limited to conversations without contents attached.