[cr141] Applies upstream commit to fix windows component build.
Chromium change:

https://source.chromium.org/chromium/chromium/src/+/e96d3a1622fdc1d98df65875499377f31699d17f

[webui browser] Remove unneeded export, workarounds from SidePanelUIBase

In a previous CL, when we added SidePanelUIBase, missing method
definition linker errors caused us to mark the class with an export
macro. In subsequent CLs, however, this led to more subtle import/
export linker errors on Windows component builds. Workarounds using
NOINLINE and trivial pass-throughs seemed to fix the errors, but then
the errors would reappear in some form in a CL further along the
dependency chain.

After removing the initial export macro and subsequent workarounds,
it seems that the missing method errors were transient (and since
they were local I do not have a record of them on the bots). Hence
I think it may be safe to remove this code.

Bug: 439850831,439850993,393144165