Set ZOrderLevel::kSecuritySurface to parent widget of PermissionPromptBubble
Previously, we only set kSecuritySurface to the prompt widget.
But this caused the prompt to be floating without its parent window,
which may feel disjointed from the overall UI.

thus, this change sets kSecuritySurface to the parent widget of the prompt.
And when the prompt bubble is closed, the parent widget's ZOrderLevel is
reset to its original value.