Block Lottie themed images for credit card / password / address save dialogs
We want to block the various header images show on the credit card, password,
and address save dialogs. Since those functions are spread across various
classes, we block the images via an override for the upstream helper function
that retrieves them.

The set of blocked images are initialized at startup, so we can avoid the need
to explicitly reference the blocked resource IDs by name in the helper which
would cause a layering violation (they're defined in
"chrome/grit/browser_resources.h")/