Remove top sites suggestions provider (#31073)
The list was not updated nor maintained, and even had invalid entries.

On device which uses a local ML model was tied to this preference for some reason too, so that keeps the preference around and has a checkbox for just that now.

For Android, it doesn't support on device suggestions because the ML
models aren't available there, so no checkbox was added there.