Run network audit basic test with default and Google search engines.
Currently we run the network audit test with the default search engine
(which is Brave), but some upstream code behaves differently when Google
is the default search engine. This change parametrizes the basic network
audit test to run one time with the default search engine and one more
time with Google as the default search engine.