Fixed gn check error brave_layout_constants.h/.cc
ERROR at //brave/browser/ui/brave_layout_constants.h:11:11: Include not allowed.
          ^-----------------------------------
It is not in any dependency of
  //brave/browser/ui:brave_layout_constants
The include file is in the target(s):
  //chrome/browser/ui:layout_constants
which should somehow be reachable.
This might be a false alarm if you are using conditional include. See "gn help nogncheck".
___________________
ERROR at //brave/browser/ui/brave_layout_constants.cc:10:11: Include not allowed.
          ^-----------------------------------
It is not in any dependency of
  //brave/browser/ui:brave_layout_constants
The include file is in the target(s):
  //chrome/browser/ui:layout_constants
which should somehow be reachable.
This might be a false alarm if you are using conditional include. See "gn help nogncheck".