Bring back RenderWidgetHostDelegate::PreHandleMouseEvent()
This method was removed from upstream with below changes because no
one uses that api. However we rely on that method for sidebar's
"on mouseover" option. This methods give a chance to handle mouse
event from browser before it's consumed by renderer.
With this method, we can show sidebar UI when mouse moves around
left/right edge of browser window. Without this api, we have to
use invisible widget to detect mouse event. Controlling that invisible
widget over the web contents is complex and buggy.
With this api, it's very simplified.

https://chromium-review.googlesource.com/c/chromium/src/+/6880411
    Remove unused PreHandleMouseEvent

    This method on Browser, BrowserWindow, WebContentsDelegate etc always
    returned false in all implementations.
    Rather than implement it for WebUIBrowserWindow (it was generating
    *a lot* of NOTIMPLEMENTED() debug messages) I figure removing it is
    better for overall code health.