Don't reset custom title when it's initial navigation
When a tab is newly created, e.g. restoring tabs, it has not yet
committed a real page. In this case the origin could be empty but we
should not reset the custom title. As the real page will be committed
soon and the custom title will be reset or persisted based on the
origin of the real page once it's committed.