Simplify macOS packaging GN actions
The previous code was there to retain backwards-compatibility with
devops code that still expected the output files at the legacy paths.
Now, the devops code has been updated to use the new paths. It is
therefore no longer necessary to maintain copies of the output files at
the legacy paths.