[Android] Update light/dark colours from Nala XML definitions
This change has been produced using a Ruby script that implements the
following logic:

1. For each <color> tag in the destination, try to find one in the
   source with the same name
2. If found, update the destination tag with the colour from the source
   tag
3. If not found, do nothing and move onto the next tag.
4. Tags that exist in the source but not the destination are ignored.

Currently, some tags exist in the destination XML that are missing from
the source and also vice versa.

Once this has been resolved, we should be able to delete the files in
this diff and move to simply copying the generated XML files straight
out of brave/leo.

Additionally, a small change has been made to the Brave Sync "Add New
Device" button to use a Nala-defined colour rather than being hard-coded
to white.

Closes brave/brave-browser#49143