[Android] Attempt to fix cut-off background image after app restart
Occasionally, users have reported that the NTP (New Tab Page) background
image is partially cut-off after restarting the app. The best
working-theory that I have around this is that there is some sort of
layout inconsistency happening when the layout is internally saved and
restored.

It is entirely possible that this problem is caused by the XML layout
definition causing some sort of inconsistency when Android decides how
to size elements - the most reasonable explanation being that when
restoring the view instead of drawing it from scratch, elements are
evaluated in a different order.

This commit attempts to programatically compute and apply the size to
the background image after built-in sizing has occured. There is quite
possibly a better, cleaner fix that could possibly be done entirely in
XML.