[Android] Support import of passwords from CSV (#31056)
* [Android] Refactor PasswordSettings.java to use an XML layout file.

The PasswordSettings.java file now integrates with a new XML file that
contains the layout for the page. This modification was completed using
Cursor but does not appear to have introduced any breaks.

There should be no semantic changes caused by this commit and the Brave
Password settings screen should continue to function exactly as it did
prior to this commit.

* [Android] Introduce new Import/Export items to password manager menu.

The password manager dialog now features two items for export and
import. Prior to this commit, Export was hidden away in a small
drop-down menu accessible from the top-right of the screen.

The import item is also present, but non-functional in this commit.

* [Android] Implement importing passwords from CSV.

We now support importing a CSV file containing the user's passwords.
This commit builds upon its parent by implementing the necessary code to
call Chromium's password manager which is responsible for executing the
import.

Additionally, the icons in the password manager screen have been changed
to new ones taken from Google's Material UI icon set.

Localisation strings have been introduced - some of which are copied
from Google's base since, after discussion, this is the way we currently
make strings usable for Android. If this changes in future, these could
be de-duplicated. Translations into other languages are still necessary
since only English is implemented.

Closes brave/brave-browser#35729