[iOS] Replace "Search Web" web view menu item with Brave's
This change replaces the "Search Web" action that appears when you've selected text on the page and opened the context menu with the "Search with Brave" action we typically placed at the end of the list. We do this because "Search Web" would use Safari or–in the case where Brave is the default–Safari's selected search engine to perform the search.