Fixed dangling ptr from NewTabPageHandler (#31420)
Resolves brave/brave-browser#49626

Cached TabInterface was dangled.
In WebAppTabStripBrowserTest.ReparentingPinsHomeTab, another url is loaded in tabbed browser's NTP.
And that tab's web contents is reparented to web app window.
When that tab is reparented, its TabModel(TabInterface) is destroyed but still NewTabPageHandler cached it.
Fixed by caching WebContents instead of TabInterface because WebContents is not invalidated duing the
web contents reparenting.