Temporary workaround for crash in PIP window.
Temporary overrides VideoOverlayWindowViews::IsTrustedForMediaPlayback
becase the call to media_session->GetRoutedFrame() in that method causes
a crash. The cause for the crash is not clear ATM. Looks like a bad the
media_session has a bad vtable because the call ends up in the
GetMediaSessionInfoSync method of media_session instead of the
GetRoutedFrame method.