Integrate AIChat with the Chromium ActorKeyedService (#31224)
* Integrate AIChat with the Chromium ActorKeyedService

This introduces the connection for providing Tools which use the ActorKeyedService to AIChat conversations. It fulfills a per-conversation owner of an Actor Task, which will be available to all Tools which will also be owned by the same ToolProvider. We provide this functionality to the Tools by means of a TaskProvider so that we won't need to pass the ContentAgentToolProvider definition to each Tool and can limit what we expose in that circular relationship.
- Introduce ContentAgentToolProvider which implements AIChat's ToolProvider. It will be used to group a set of Tools for content agent purposes and is only created for the agent profile
- ContentAgentToolProvider will be responsible for using the Actor Service to create and own Tasks and the Tabs within the Tasks.
- Add 1 tool in order to provide context. The rest will follow in subsequent PR.
- Overrides Chromium's actor site policy so that the tool can act on and with any URL. We change the fallback to allow instead of deny. We restrict its use to Tabs created within the conversation's task.

See the integration branch for more context and a working version:
- https://github.com/brave/brave-core/pull/30743

* don't need site_policy override anymore

* more cr141 fixes

* review feedback

* review feedback yrilou

* browser tools readme