[AI Chat] [Content Agent] Assistant Detail Tool
A simple tool for the LLM to save information it deems important before the context is truncated due to the size of page content results for previous content agent tool use requests. The tool itself doesn't do anything since the content itself is stored in the conversation history and context provided back to the LLM simply by the LLM submitting the data as tool input.
For the UI we simply show the label that the AI is storing information. When clicked, the information is shown and is streamed in if the tool use request is still being received in chunks.