[ads] Only allow conversion confirmations for Search Result ads
Determines if a transaction is allowed to send a confirmation based on
its ad type and confirmation type. It returns true for all confirmation
types when the ad type is Notification, New Tab Page, Promoted Content,
or Inline Content. For Search Result ads, it only allows sending a
confirmation if the confirmation type is kConversion; otherwise, it
returns false.