[cr142] Fix timing issue in `BraveWindowNameBrowserTest.SameOrigin` browser test
`BraveWindowNameBrowserTest.SameOrigin` loads a document that rus a script that
sets `window.location.href` to a new URL. Since the script is in `<head>` it may
run before the initial navigation is fully committed. Blink therefore saw the
previous document in a pre-commit state where its `SecurityOrigin` was still
opaque. When it compared origins, they didn't match due to that perceived
difference and `window.name` was erroneously cleared.

Instead of this, we set `location.href` once the document has finished loading.
Then the same-origin check won't fail.