Fix intermittent browser tests BatAds failure
This PR ensures that Brave Ads mojom active callbacks are invoked
with failure arguments when the Brave Ads service mojom connection
is closed. By doing so, it resolves intermittent DCHECKs that can
occur during browser shutdown, particularly when shutdown is
triggered immediately after startup, which is a scenario often
encountered in browser tests.