[AI Chat] [Content Agent] Introduce other content agent tools (#30743)
This introduces the rest of the tools for manipulating content via the Actor service (click, drag, scroll, select, type, navigate history).

In #31224 we introduced a single content agent tool (the navigation tool) and in this PR we repeat the same pattern for the rest of the tools.

These tools commonly define "target" parameters - elements on which to act. See target_util.h, target_util_unittest.cc and target_test_util.h for all the common functionality around "target".