Add an Email Aliases cr-link-row in the autofill settings page (#31717)
* Add an Email Aliases cr-link-row in the autofill settings page

The autofill and passwords settings page has links for password
manager, payment methods, addresses and more. This change adds
an additional link that takes users to the Email Aliases page

Includes browser tests to confirm that, the link only appears
when the Email Aliases flag is enabled, and that clicking on the
link takes the user to the page at brave://settings/email-aliases

Also: implements getAssociatedControlFor and removes search
params on click.