Force context menu on Shift + Right Click on elements in pages
This PR introduces a feature that allows users to open the context menu
even if `event.preventDefault()` is called on the `contextmenu` event
when the Shift key is pressed.

This is particularly useful for users who want to access the context
menu on web pages that disable it through JavaScript.

This is achieved by bypassing the `preventDefault` check for
`contextmenu` events when the Shift key is detected as being pressed in
the "EventDispatcher" so that Node's `DefaultEventHandler` is called.