[Midnight] Don't apply opacity to inactive tab foreground in midnight theme (#31762)
Don't apply opacity to inactive tab foreground in midnight theme

When the tab is inactive in midnight theme, the foreground color should not
be applied with opacity. This patch fixes the issue by moving the logic to
BraveTabColorMixer BraveTabStyleViews.

Also, fixes bugs that foreground color is wrong when frame is inactive.