[AI Chat] [Content Agent] Make decision when to show a group of Assistant conversation entries as a Task
The Task UI is stubbed and will be introduced in a follow-up PR.
A Task is defined as where there is a tool loop for a response (multiple assistant entries in sequence) *and* it does not use a non-task tool *and* there is at least one text event within that tool loop. We might change this to only consider task tools instead of non-task tools, depending on which list is easier to maintain.
For example this is a Task:
```
-- First Assistant turn
- Completion text: "I am going to run that operation now"
- Tool use: 'a-task-tool-1'
-- Second Assistant turn
- Tool use: 'a-task-tool-2'
```

And this is a Task:
```
-- First Assistant turn
- Tool use: 'a-task-tool-1'
- Tool use: 'a-task-tool-2'
-- Second Assistant turn
- Completion text: "I have the answer for you: 42"
```

But this is *not* a Task (no completion event):
```
-- First Assistant turn
- Tool use: 'a-task-tool-1'
-- Second Assistant turn
- Tool use: 'a-task-tool-2'
```

And nor is this (no task tool):
```
-- First Assistant turn
- Completion text: "I am going to run that operation now"
- Tool use: 'NON-task-tool-1'
-- Second Assistant turn
- Completion text: "I have the answer for you: 42"
```

Or this (only 1 task tool):
```
-- First Assistant turn
- Completion text: "I am going to run that operation now"
- Tool use: 'task-tool-1'
-- Second Assistant turn
- Completion text: "I have the answer for you: 42"
```