[cr143] GN direct inclusion of brave gni source
This change corrects two patch files that had previously relied on
`brave/build/config/brave_build.gni` being pulled in for the whole build
to make a reference to a particular constant. It is better to just
add the import statement with the patched line directly, as that removes
one point of failure.