[wallet][CodeHealth] Simplify test-only code for `PasswordEncryptor`
This PR simplifies the way how test-only code to inject the salt and
nonce is inject. In particular, this change moves away from having
`base::NoDestructor` instances, and actually provides only global
pointers that the tests are supposed to manage. This is more efficient
memory-wise, and also performance-wise, as `base::RepeatingCallback` is
a refcounted object.

This change also makes this customisation aspect a discreet
implementation detail to the `KeyringService`, as this customisation at
the moment is only needed for this particular class tests.

Branching is done in a way that `CHECK_IN_TEST` guarantees there is no
possibility that production code is running through test branches.

`GetPbkdf2IterationsForTesting` has be removed completely, as this
particular customisation wasn't in use at all.

Finally, this change corrects the signature of pref function to take a
`std::string_view`. This avoids silent string allocations, which ends up
being common with these functions as the key tends to be a `constexpr`
value.

A lot of code has been moved from `keyring_service_prefs.cc` and into
`keyring_service.cc`, as recommended by the style guide that internal
linkage should be preferred for functions that can be declared as such.

Resolves https://github.com/brave/brave-browser/issues/50271