[Leo] Fix rewrite-in-place via context menu
It was regressed when we migrate to conversation API, on server side we
missed adding UserText event and it's associated prompt, on client side
we are sending a RewriteEvent which is a base event of all rewrite
events, but server is expecting a specific type of rewrite event like
RequestImproveEvent or a RequestChangeTone event with a tone field.

Changes in this PR:
- Add missing rewrite events and use specific rewrite event for each
  rewrite action
- Provide tone field for change tone events
- Properly support delta responses when replacing which is now used in
  all engines
- Add missing seed and stop sequence in
  EngineConsumerOAIRemote::GenerateRewriteSuggestion for fixing BYOM
  support, and remove unneccesary ending tag matching in
  BindParseRewriteReceivedData

There will be a seperate PR on server side to add UserText event to
fix this issue together.