[wallet][safe-numerics] Improve `BtcLikeSerializerStream` use
This PR greatly simplifies `BtcLikeSerializerStream`, but making the
vector used as a buffer part of the class itself. This lines up with how
all uses of this class occur, with the vector created just before the
serialiser, so making the vector a member of the serialiser simplifies
things significantly.

A few `checked_cast`s have been added to avoid future mistakes around
this particular area of the code.

Bug https://github.com/brave/brave-browser/issues/50271