[Android] Android widget scroll (uplift to 1.84.x) (#31841)
* Android widget scroll (#31635)

 [Android] Makes a scroll in quick action search widget

There are still cases that row with tiles is cut or not displayed when
there is space for them. We better show all fav tiles when the widget
is not in a minimum height and use scroll there.
Resolves: https://github.com/brave/brave-browser/issues/49890

* [Android] Move BookmarkWidgetService to base module (#31675)

Currently we have a failure on aab builds:
Service BookmarkWidgetService should be declared in the base manifest.
That commit moves it to base module and renames to
QuickActionSearchAndBookmarkWidgetService.
Resolves: https://github.com/brave/brave-browser/issues/49959