[AI Chat] [Content Agent] Display a thumbnail of a tab that is controlled by a task's set of tools
Completes the sub-feature of displaying the tab thumbnail by building on the following PR which provided the tab handle (as ID int) to the UI:
- https://github.com/brave/brave-core/pull/31945
  - This PR also contained the test to verify that the tab ID is being passed from that ToolProvider task to the UI client.

This adds the existing ThumbnailTracker to the WebUI class which powers the "untrusted" conversation content frame. This class is desktop only and is guarded in Chromium by `IS_ANDROID`, so the same guards are used here.

Whilst the thumbnail trackking and passing to the UI is used for the content agent conversation only here - it is not tied to it. We could conceivably use it for previews of regularly attached Tabs. That's why I haven't used the `enable_brave_ai_chat_agent_profile` build flag.