[Android] Fix VPN Enrollment screens not closing
Previously, we removed a call to bringChromeTabbedActivityToTheTop in
order to prevent closing the server country selection dialog.

However, this introduced a regression in that the enrollment screens
which depend on this call no longer close and the user would resultantly
have to back out of them.

Now, we conditionally invoke the call unless the activity is coming from
either the country selection or city selection dialog.

Unit tests have been added to verify the behaviour.

Resolves brave/brave-browser#50163